Added support for assisted transparency
By assisted, I mean that nstwm does not support transparency itself,
but if for example xcompmgr is running, providing the real
transparency support, we will cooperate and not break it.
We must make our decorations use a visual that has an alpha channel,
even if we're not going to use it ourselves, or we'll prevent the
contained window from using transparency.